// src/redux/types/tableOrderTypes.ts
export enum TableOrderActionTypes {
  FETCH_TABLE_ORDER = 'FETCH_TABLE_ORDER',
  FETCH_TABLE_ORDER_SUCCESS = 'FETCH_TABLE_ORDER_SUCCESS',
  FETCH_TABLE_ORDER_ERROR = 'FETCH_TABLE_ORDER_ERROR',
  SET_TABLE_ORDER_PAGE = 'SET_TABLE_ORDER_PAGE',
  SET_TABLE_ORDER_PERIOD = 'SET_TABLE_ORDER_PERIOD',
  SELECT_PERIOD = 'SELECT_PERIOD',
  // SELECT_MODEL = 'SELECT_MODEL',
  // SELECT_CITY = 'SELECT_CITY',
  // SELECT_STATUS = 'SELECT_STATUS',
}